Internetveikals/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.2
django-allauth==0.51.0
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-js-asset==2.0.0
django-modeltranslation==0.18.4
django-mptt==0.14.0
django-reset-migrations==0.4.0
django-s3-sqlite==0.0.5
django-un-foo==0.1.9
idna==3.4
mysqlclient==2.1.1
oauthlib==3.2.1
Pillow==9.2.0
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
types-cryptography==3.3.23
tzdata==2022.4
urllib3==1.26.12
six
